moved HandlerFactory and RouterFactory to router-impl module
